Fix all clippy warnings across workspace
Resolve collapsible_if, redundant closures, unnecessary borrows,
literal format strings, unnecessary casts, contains_key+insert,
map_or simplification, &PathBuf→&Path, missing Default impls,
and complex type warnings. Add clippy rule to CLAUDE.md.

+++ crates/sandogasa-bugzilla/src/client.rs
@@ -70,7 +70,7 @@ impl BzClient {
 
         loop {
             let limit = if max_results > 0 {
-                PAGE_SIZE.min(max_results - offset as u64)
+                PAGE_SIZE.min(max_results - offset)
             } else {
                 PAGE_SIZE
             };
